backward unsplit one file
