+++ config/splat.pspeu.dra.yaml
@@ -92,7 +92,6 @@ segments:
       - [0x23FE0, c, dra_psp/23FE0] # O0
       - [0x26948, c, dra_psp/26948]
       - [0x29420, c, dra_psp/29420]
-      - [0x29760, c, dra_psp/29760]
       - [0x2BF60, c, dra/71830]
       - [0x35238, c, dra_psp/35238]
       - [0x353B0, c, dra/704D0] # O0
@@ -173,7 +172,7 @@ segments:
       - [0xEE628, .rodata, dra_psp/1DDC0]
       - [0xEE680, .rodata, dra_psp/23FE0]
       - [0xEE6C8, .rodata, dra_psp/26948]
-      - [0xEE708, .rodata, dra_psp/29760]
+      - [0xEE708, .rodata, dra_psp/29420]
       - [0xEE7B0, .rodata, dra/71830]
       - [0xEE840, .rodata, dra/704D0]
       - [0xEE898, .rodata, dra/84B88]
